♻️: folder and function names refactor
