+++ z-economy/src/app/components/forms/AddCategory/AddCategoryForm.tsx
@@ -7,8 +7,8 @@ import { Tooltip } from 'react-tooltip';
 import { Signal } from '@preact/signals-react';
 import { useOutsideClick } from '@utils/mouseUtils';
 import { v4 as uuidv4 } from 'uuid';
-import { Category } from '@core/budget/budget/domain/Category';
-import { SubCategory } from '@core/budget/budget/domain/SubCategory';
+import { Category } from '@core/budget/category/domain/Category';
+import { SubCategory } from '@core/budget/category/domain/SubCategory';
 
 interface AddCategoryFormProperties {
   isOpen: Signal<boolean>;
