+++ z-economy/src/app/components/molecules/CategoryTableButtons/CategoryTableButtons.tsx
@@ -4,7 +4,7 @@ import styles from './CategoryTableButtons.module.scss';
 import { Typography } from '@atoms/Typography/Typography';
 import { AddCategoryForm } from '../../forms/AddCategory/AddCategoryForm';
 import { useSignal } from '@preact/signals-react';
-import { Category } from '@core/budget/budget/domain/Category';
+import { Category } from '@core/budget/category/domain/Category';
 
 interface CategoryTableButtonsProperties {
   createCategoryGroup: (c: Category) => Promise<void>;
