+++ z-economy/src/core/budget/category/domain/CategoryRepository.ts
@@ -1,5 +1,5 @@
-import { Category } from '@core/budget/budget/domain/Category';
-import { SubCategory } from '@core/budget/budget/domain/SubCategory';
+import { Category } from '@core/budget/category/domain/Category';
+import { SubCategory } from '@core/budget/category/domain/SubCategory';
 
 export interface CategoryRepository {
   getAll(month: string, year: string): Promise<Category[]>;
